LayoutTests/imported/w3c/web-platform-tests/css/css-contain/contain-size-grid-003-expected.txt


PASS .grid 1
PASS .grid 2
PASS .grid 3
PASS .grid 4
PASS .grid 5
PASS .grid 6
PASS .grid 7
PASS .grid 8
PASS .grid 9
PASS .grid 10
PASS .grid 11
PASS .grid 12
PASS .grid 13
PASS .grid 14
PASS .grid 15
PASS .grid 16
PASS .grid 17
PASS .grid 18
PASS .grid 19
PASS .grid 20
PASS .grid 21
PASS .grid 22
PASS .grid 23
PASS .grid 24
PASS .grid 25
PASS .grid 26
PASS .grid 27
PASS .grid 28
PASS .grid 29
PASS .grid 30
PASS .grid 31
PASS .grid 32
PASS .grid 33
PASS .grid 34
PASS .grid 35
PASS .grid 36
PASS .grid 37
FAIL .grid 38 assert_equals:
<div class="grid" style="grid: repeat(auto-fit, 75px) / repeat(auto-fit, 100px);" data-expected-width="0" data-expected-height="0">
    <div></div>
  </div>
width expected 0 but got 100
FAIL .grid 39 assert_equals:
<div class="grid" style="grid: auto 50px / 100px auto;" data-expected-width="100" data-expected-height="50">
  <div data-expected-width="100" data-expected-height="10" data-offset-x="0" data-offset-y="0">X</div>
  <div data-expected-width="40" data-expected-height="10" data-offset-x="100" data-offset-y="0">XX</div>
  <div data-expected-width="100" data-expected-height="50" data-offset-x="0" data-offset-y="10">XXX</div>
  <div data-expected-width="40" data-expected-height="50" data-offset-x="100" data-offset-y="10">XXXX</div>
</div>
width expected 100 but got 140

X
XX
XXX
XXXX
